refactor: reduce duplication, add test
